Added a way to sort the admin components with an index
