+++ Moonlight/Core/CoreFeature.cs
@@ -7,9 +7,10 @@ using MoonCoreUI.Services;
 using Moonlight.Core.Configuration;
 using Moonlight.Core.Database;
 using Moonlight.Core.Database.Entities;
-using Moonlight.Core.Implementations.AdminColumns;
 using Moonlight.Core.Implementations.Diagnose;
+using Moonlight.Core.Implementations.UI.Admin.AdminColumns;
 using Moonlight.Core.Interfaces;
+using Moonlight.Core.Interfaces.Ui.Admin;
 using Moonlight.Core.Models;
 using Moonlight.Core.Models.Abstractions;
 using Moonlight.Core.Models.Abstractions.Feature;
