--- Moonlight/Core/Interfaces/IAdminDashboardColumn.cs
@@ -1,8 +1,0 @@
-﻿using Microsoft.AspNetCore.Components;
-
-namespace Moonlight.Core.Interfaces;
-
-public interface IAdminDashboardColumn
-{
-    public Task<RenderFragment> Get();
-}
