--- Moonlight/Core/Interfaces/IAdminDashboardComponent.cs
@@ -1,8 +1,0 @@
-﻿using Microsoft.AspNetCore.Components;
-
-namespace Moonlight.Core.Interfaces;
-
-public interface IAdminDashboardComponent
-{
-    public Task<RenderFragment> Get();
-}
