+++ Moonlight/Core/Interfaces/UI/Admin/IAdminDashboardColumn.cs
@@ -1,0 +1,8 @@
+﻿using Moonlight.Core.Models.Abstractions;
+
+namespace Moonlight.Core.Interfaces.Ui.Admin;
+
+public interface IAdminDashboardColumn
+{
+    public Task<UiComponent> Get();
+}
