+++ Moonlight/Core/Interfaces/UI/Admin/IAdminDashboardComponent.cs
@@ -1,0 +1,8 @@
+﻿using Moonlight.Core.Models.Abstractions;
+
+namespace Moonlight.Core.Interfaces.Ui.Admin;
+
+public interface IAdminDashboardComponent
+{
+    public Task<UiComponent> Get();
+}
