+++ Moonlight/Core/Models/Abstractions/UiComponent.cs
@@ -1,0 +1,10 @@
+﻿using Microsoft.AspNetCore.Components;
+
+namespace Moonlight.Core.Models.Abstractions;
+
+public class UiComponent
+{
+    public required RenderFragment Component { get; set; }
+
+    public int Index { get; set; } = 0;
+}
