+++ Moonlight/Features/Servers/ServersFeature.cs
@@ -2,13 +2,14 @@ using MoonCore.Helpers;
 using MoonCore.Services;
 using Moonlight.Core.Configuration;
 using Moonlight.Core.Interfaces;
+using Moonlight.Core.Interfaces.Ui.Admin;
 using Moonlight.Core.Models.Abstractions.Feature;
 using Moonlight.Core.Services;
 using Moonlight.Features.Servers.Actions;
 using Moonlight.Features.Servers.Configuration;
 using Moonlight.Features.Servers.Http.Middleware;
-using Moonlight.Features.Servers.Implementations.AdminColumns;
 using Moonlight.Features.Servers.Implementations.Diagnose;
+using Moonlight.Features.Servers.Implementations.UI.Admin.AdminColumns;
 using Moonlight.Features.Servers.Models.Enums;
 using Moonlight.Features.Servers.Services;
 using Moonlight.Features.Servers.UI.Components.Cards;
