+++ Moonlight/Moonlight.csproj
@@ -53,7 +53,7 @@
       <Folder Include="Core\Database\Migrations\" />
       <Folder Include="Core\Http\Requests\" />
       <Folder Include="Core\Http\Resources\" />
-      <Folder Include="Core\Implementations\AdminComponents\" />
+      <Folder Include="Core\Implementations\UI\Admin\AdminComponents\" />
       <Folder Include="Core\UI\Components\Forms\" />
       <Folder Include="Features\Dummy\Configuration\" />
       <Folder Include="Features\Dummy\Entities\" />
@@ -75,6 +75,7 @@
       <Folder Include="Features\FileManager\Http\Requests\" />
       <Folder Include="Features\FileManager\Http\Resources\" />
       <Folder Include="Features\Servers\Http\Resources\" />
+      <Folder Include="Features\Servers\Implementations\UI\Admin\" />
       <Folder Include="storage\" />
       <Folder Include="Styles\" />
       <Folder Include="wwwroot\css\" />
